getPrinted done

+++ src/main/java/iovi/dispatcher/PrintDispatcher.java
@@ -23,7 +23,7 @@ public interface PrintDispatcher {
     List<Document> getPrinted(SortingType sortingType);
 
     /**Получение среднего времени печати в мс. Возвращает 0, если ни один документ не напечатан*/
-    long getAveragePrintDuration();
+    int getAveragePrintDuration();
 
 
 }
